move nat/redirect logic completely to nat_rewrite
Previously, Nat_lookup.insert took a "mode" argument to define whether
it should make entries corresponding to a static port mapping or a
conventional NAT entry out of a pair of four (Ipaddr.t, port) mappings.
Now, Nat_lookup.insert takes a set of mappings (along with a protocol
number) directly, and formulating the (internal_lookup, external_lookup,
internal_mapping, external_mapping) set correctly is the responsibility
of the caller.  The logic for doing this correctly in the case of Nat
and Redirect is now in nat_rewrite.

This makes nat_lookup.ml a bit less magic at the expense of making
nat_rewrite.ml a bit messier; further cleanup of this logic is probably
warranted.

Additionally, test_lookup.ml contained some tests for consistent behavior
in the cases of Nat vs Redirect inserts that should move to
test_rewrite in a future commit.