Store keys and values directly, rather than in a map.
Previously mirage-nat would construct a Map of translations and store
this map in Irmin on a per-protocol basis.  This meant any operations on
the table required fetching and reading the map from Irmin, then
doing a lookup on the specific key of interest; moreover changes
required re-serialization of the entire table.

This commit changes the
strategy of mirage-nat to store keys and values directly in Irmin with a
custom Path and corresponding Step.  The Path does not actually
implement a hierarchy; while such a structure might be useful for
organizing this data generally (e.g.,
protocol/source_port/destination_port/source_ip/destination_ip as a
path), the interface specified in Irmin.Path.S isn't a good fit for it
at the moment.

Storing the data directly allows the easy use of watches; at the
previous granularity they served only as a notification that some change
had been made to the datastore, requiring the user to dig around in the
map themselves.