rework handling of ICMP error messages
Previously, we discarded most ICMP error messages because the
encapsulated IPv4 packet appeared to be incorrectly truncated, as we
were using the `tcpip` library's IPv4 header unmarshaller,
which checked the length field of IPv4 packets against the length
of the payload, and would return an error if they did not match.  This
isn't appropriate for the ICMP error case, where the message is
truncated.  `tcpip` now exposes a function for just parsing the header,
which we use here to get the header information and correctly process
the error packet.

There are also some style changes which clarify handling for this nested
case where ICMP error messages encapsulate an IPv4 packet, and
comments on how this case should be handled in a NAT device and why.

Co-authored-by: linse <linse@linse.me>